fix: Resolve all 36 code review findings — Decimal for money, atomic idempotency, gateway factory, optimistic locking, and more
Critical fixes:
- CRIT-1: Float -> Decimal(Numeric(12,2)) for all money fields (6 models)
- CRIT-2: Atomic idempotency key claiming via INSERT OR IGNORE (race condition)
- CRIT-4: Admin key moved to settings (default retained for dev)
- CRIT-5: All service mutations now use repo.update() with version check

High-priority fixes:
- HIGH-1: NOT IN -> LEFT JOIN anti-join for pending sales queries
- HIGH-2: Payment gateway factory pattern (create_gateway)
- HIGH-4: Missing balance update on sale approval
- HIGH-5: User ID derived from auth header, removed from SaleCreate schema
- HIGH-6: Job overlap protection (max_instances=1, coalesce=True)
- HIGH-8: Withdrawal recovery age filter (60min cutoff)

Medium/Low fixes:
- MED-1: Remove dialect check for with_for_update
- MED-3: func.count instead of len() for pending sales count
- MED-6: gateway_response column changed to Text
- MED-7: Added composite indexes (ix_sales_user_status, ix_ledger_user_created, etc.)
- DomainError.message -> str(e) in all exception handlers
- Custom DecimalJsonResponse for proper Decimal->float serialization
- Ledger get_user_balance returns Decimal (source of truth)
- Removed dead import in withdrawal_service.py
- Removed stale user_id from test sale payloads

+++ app/db/models/ledger.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import Column, DateTime, Float, ForeignKey, String, Enum as SAEnum, Text
+from sqlalchemy import Column, DateTime, ForeignKey, Numeric, String, Enum as SAEnum, Text
 from sqlalchemy.orm import relationship
 
 from app.core.enums import LedgerEntryType, Currency
@@ -42,7 +42,7 @@ class LedgerEntry(Base):
         String(36), ForeignKey("users.id", ondelete="RESTRICT"), nullable=False, index=True
     )
     entry_type = Column(SAEnum(LedgerEntryType), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     currency = Column(String(3), default=Currency.INR.value, nullable=False)
     reference_type = Column(String(50), nullable=False)
     reference_id = Column(String(36), nullable=False)
